add relation between Prof module and student && add list inscription …

+++ backend/database/migrations/2024_05_09_121520_create_students_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('students', function (Blueprint $table) {
-            $table->integer("apogee")->primary()->autoIncrement();
+            $table->unsignedBigInteger("apogee")->primary()->autoIncrement();
             $table->string("password");
             $table->string("phone_number");
             $table->string("gender");
